Removed new field and migrations
