+++ validator/forms/__init__.py
@@ -3,4 +3,5 @@ from .data_configuration import *
 from .validation_run import *
 from .signup import *
 from .user_profile import *
+from .results_sorting import *
 from .publishing import *
