+++ validator/forms/custom_widgets.py
@@ -95,83 +95,3 @@ class YearChoiceField(forms.fields.ChoiceField):
             v_date = datetime(year=int(v), month=12, day=31, hour=23, minute=59, second=59, tzinfo=UTC)
 
         return v_date
-
-
-class ResultsSortingForm(forms.Form):
-    """
-    Allows the user to select a key which is used for sorting, and whether it
-    should be sorted ascending or descending.
-    """
-
-    sort_key = forms.fields.ChoiceField(
-        choices=[
-            ("start_time", "Date"),
-            ("name_tag", "Name"),
-            ("progress", "Status"),
-            ("ref_dataset_name", "Reference dataset"),
-        ],
-        required=False,
-    )
-    sort_order = forms.fields.ChoiceField(
-        choices=[
-            ("desc", "descending"),
-            ("asc", "ascending"),
-        ],
-        required=False,
-    )
-
-    def __init__(self, *args, **kwargs):
-        super().__init__(*args, **kwargs)
-        # set initial values on field level too
-        for field in self.fields:
-            self.fields[field].initial = self.get_initial_for_field(
-                self.fields[field], field
-            )
-
-    def clean(self):
-        cleaned_data = super().clean()
-        # replace missing/invalid values with initial values
-        for field in self.fields:
-            if field not in self.data or field not in cleaned_data:
-                cleaned_data[field] = self.get_initial_for_field(
-                    self.fields[field], field
-                )
-
-    @classmethod
-    def get_sorting(
-            cls,
-            request,
-            initial_key="start_time",
-            initial_order="desc"
-    ):
-        """
-        Creates a sorting form from the given request and returns it together
-        with the sorting key/order for use in ``order_by``.
-
-        Parameters
-        ----------
-        request : request
-        initial_key : str, optional
-            Options are: "start_time" (default), "name_tag", "progress",
-            "ref_dataset_name".
-        initial_order : str, optional
-            Options are: "desc" (default), "asc"
-
-        Returns
-        -------
-        sorting_form : ResultsSortingForm
-        order : str
-            String for use in ``order_by``, including key and ordering
-            direction.
-        """
-        key, order = initial_key, initial_order
-        form = cls(
-            request.GET,
-            initial={"sort_key": initial_key, "sort_order": initial_order}
-        )
-        if form.is_valid():
-            key = form.cleaned_data["sort_key"]
-            order = form.cleaned_data["sort_order"]
-        order_string = {"asc": "", "desc": "-"}[order] + key
-
-        return form, order_string
