+++ validator/models/validation_run.py
@@ -61,7 +61,6 @@ class ValidationRun(models.Model):
     progress = models.IntegerField(default=0)
 
     reference_configuration = models.ForeignKey(to=DatasetConfiguration, on_delete=models.SET_NULL, related_name='ref_validation_run', null=True)
-    ref_dataset_name = models.CharField(max_length=30)
 
     scaling_ref = models.ForeignKey(to=DatasetConfiguration, on_delete=models.SET_NULL, related_name='scaling_ref_validation_run', null=True)
     scaling_method = models.CharField(max_length=20, choices=SCALING_METHODS, default=MEAN_STD)
