+++ validator/views/published_results.py
@@ -1,7 +1,7 @@
 from django.core.paginator import Paginator, EmptyPage, PageNotAnInteger
 from django.shortcuts import render
 
-from validator.forms.custom_widgets import ResultsSortingForm
+from validator.forms import ResultsSortingForm
 from validator.models import ValidationRun
